Minor change to fix tests

+++ tests/test_skytemp.py
@@ -10,7 +10,7 @@ def test_fileexists():
 
 def test_skytemp_withfile():
     s = SkyTemp(0, 0, "haslam408_ds_Remazeilles2014.fits")
-    assert type(s.temp408) is numpy.float64
+    assert type(s.temp408) is float
     assert round(s.temp408, 0) == 1804.
-    assert type(s.get_temp(1400)) is numpy.float64
+    assert type(s.get_temp(1400)) is float
     #print("Temp. at (0, 0) at 1400 MHz is {:.3f}".format(s.get_temp(1400)))
